+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Development
 - (Fill this out as you fix issues and develop your features).
 - Fix .only() working improperly after using .count() of the same instance of QuerySet
 - POTENTIAL BREAKING CHANGE: All result fields are now passed, including internal fields (_cls, _id) when using `QuerySet.as_pymongo` #1976
+- Fix InvalidStringData error when using modify on a BinaryField #1127
 
 =================
 Changes in 0.16.3